Do a different check for fips on immucore
It appears that immucore is not using the ssl libs anywhere so we cannot
check the final binary for the FIPS string.

Thankfully immucore has a version command that prints the version and
the go version used to compile it, so we can grep for boringcrypto in
there to know if the go version was build with boringcrypto.

We should probably use the same version command everywhere so its easier
to show the go version that our binaries were compiled with

Signed-off-by: Itxaka <itxakaserrano@gmail.com>