set kairos services to simple
If they are of type oneshot, systemd will wait until the pid exits
before markling the service as running. Instead it will mark them as
starting which is wrong and on livecd it causes the systemctl status
output to be on starting forever.

Instead lets set those services to simple, which means they will be
marked as started once the binary has been launched.

Signed-off-by: Itxaka <itxaka@kairos.io>