Create a sentinel file to avoid running the fix every time
This file can also be created by the user to skip this fix if they so
desire. For example, they may have files owned by some other user and
they don't want to mess it up. In that case they can run a similar
script themselves, but with additional logic.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>